+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ### 1.1.17
 * no longer needs the me-tools package;
 * no longer uses the `Exceptionist` class;
-* provides its own `Entity` abstract class and no longer uses the one provided by `php-tools`.
+* provides its own `Entity` abstract class and no longer uses the one provided by `php-tools`;
+* ready for `php-tools` 1.8.
 
 ### 1.1.16
 * added the `LINK_SCANNER_TMP` constant. This constant, which can be defined before the plugin, affects all used files;